cambios a war
