+++ src/main/java/com/enterprise/usco/uvscannerservidor/UvScannerServidorAplicacion.java
@@ -8,14 +8,22 @@ import org.hibernate.ejb.HibernateEntityManagerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.support.SpringBootServletInitializer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
 @EnableAutoConfiguration
 @Import(WebSecurityConfig.class)
-public class UvScannerServidorAplicacion {
+public class UvScannerServidorAplicacion 
+       extends SpringBootServletInitializer {
 
+@Override
+protected SpringApplicationBuilder configure(SpringApplicationBuilder builder) {
+return builder.sources(new Class[]{UvScannerServidorAplicacion.class}); //To change body of generated methods, choose Tools | Templates.
+}
+    
     public static void main(String[] args) {
         SpringApplication.run(UvScannerServidorAplicacion.class, args);
     }
